Rename `ConfigurationFormat` to `Format`
