+++ codegen/java/src/test/kotlin/io/spine/protodata/codegen/java/annotation/SuppressWarningsAnnotationSpec.kt
@@ -32,7 +32,7 @@ import io.spine.protodata.backend.Pipeline
 import io.spine.protodata.codegen.java.JAVA_FILE
 import io.spine.protodata.codegen.java.WithSourceFileSet
 import io.spine.protodata.settings.Configuration
-import io.spine.protodata.settings.ConfigurationFormat.PROTO_JSON
+import io.spine.protodata.settings.Format.PROTO_JSON
 import kotlin.io.path.Path
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Nested
