+++ compiler/src/test/kotlin/io/spine/protodata/backend/CodegenContextSpec.kt
@@ -37,7 +37,7 @@ import io.spine.protodata.ConfigurationError
 import io.spine.protodata.renderer.SourceFileSet
 import io.spine.protodata.renderer.codeLine
 import io.spine.protodata.settings.Configuration
-import io.spine.protodata.settings.ConfigurationFormat
+import io.spine.protodata.settings.Format
 import io.spine.protodata.test.AnnotationInsertionPointPrinter
 import io.spine.protodata.test.CatOutOfTheBoxEmancipator
 import io.spine.protodata.test.DeletedTypeRepository
@@ -365,7 +365,7 @@ class CodegenContextSpec {
                     SourceFileSet.create(source2, destination2)
                 ),
                 request = request,
-                config = Configuration.rawValue(expectedContent, ConfigurationFormat.PLAIN)
+                config = Configuration.rawValue(expectedContent, Format.PLAIN)
             )()
 
             val firstFile = destination1.resolve(ECHO_FILE)
